Make ENIP follow the same service-based model
