+++ CMakeLists.txt
@@ -232,7 +232,7 @@ if(OPLC_MAIN_PROGRAM)
     message("Compile main program enabled")
 
     set(OPLC_USER_DIR ${PROJECT_SOURCE_DIR}/etc/src)
-    file(GLOB oplc_SRC runtime/core/*.cpp runtime/core/dnp3s/*.cpp runtime/core/modbusslave/*.cpp runtime/core/modbusmaster/*.cpp runtime/core/service/*.cpp runtime/vendor/inih-r46/*.c)
+    file(GLOB oplc_SRC runtime/core/*.cpp runtime/core/dnp3s/*.cpp runtime/core/modbusslave/*.cpp runtime/core/modbusmaster/*.cpp runtime/core/service/*.cpp runtime/core/enip/*.cpp runtime/vendor/inih-r46/*.c)
 
     include_directories(${OPLC_USER_DIR})
     include_directories(runtime/core)
