+++ runtime/core/CMakeLists.txt
@@ -26,7 +26,7 @@ endif()
 include_directories(lib)
 
 # The primary source is everything in this directory
-file(GLOB oplc_SRC *.cpp dnp3s/*.cpp service/*.cpp modbusslave/*.cpp modbusmaster/*.cpp)
+file(GLOB oplc_SRC *.cpp dnp3s/*.cpp service/*.cpp modbusslave/*.cpp modbusmaster/*.cpp enip/*.cpp)
 
 message("In runtime")
 message(${oplc_SRC})
