--- runtime/core/enip.h
@@ -1,26 +1,0 @@
-//-----------------------------------------------------------------------------
-// Copyright 2019 Thiago Alves
-// This file is part of the OpenPLC Software Stack.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http ://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissionsand
-// limitations under the License.
-
-// This file contains the structured used by enip.cpp to process
-// EtherNet/IP requests.
-// UAH, Sep 2019
-//-----------------------------------------------------------------------------
-
-/** \addtogroup openplc_runtime
- *  @{
- */
-
-/** @} */
