+++ runtime/core/service/service_registry.cpp
@@ -19,6 +19,7 @@
 #include "service_registry.h"
 #include "interactive_server.h"
 #include "pstorage.h"
+#include "enip/enip.h"
 #include "modbusslave/slave.h"
 #include "modbusmaster/master_indexed.h"
 #include "dnp3s/dnp3.h"
@@ -34,6 +35,7 @@ ServiceDefinition* services[] = {
 #ifdef OPLC_DNP3_OUTSTATION
     new ServiceDefinition("dnp3s", dnp3s_service_run),
 #endif
+    new ServiceDefinition("enip", enip_service_run)
 };
 
 ServiceDefinition* services_find(const char* name)
