+++ utils/glue_generator_src/test/glue_generator_test.cpp
@@ -29,19 +29,19 @@ using namespace std;
 #define PREFIX "void glueVars()\n{\n"
 #define POSTFIX "}\n\n"
 #define EMPTY_INPUT "/// The size of the array of input variables.\n\
-extern std::size_t const OPLCGLUE_INPUT_SIZE(0);\n\
+extern const std::size_t OPLCGLUE_INPUT_SIZE(0);\n\
 GlueVariable oplc_input_vars[] = {\n\
 };\n\n"
 #define EMPTY_OUTPUT "/// The size of the array of output variables.\n\
-extern std::size_t const OPLCGLUE_OUTPUT_SIZE(0);\n\
+extern const std::size_t OPLCGLUE_OUTPUT_SIZE(0);\n\
 GlueVariable oplc_output_vars[] = {\n\
 };\n\n"
 #define EMPTY_INPUT_BOOL "/// Size of the array of input boolean variables.\n\
-extern std::size_t const OPLCGLUE_INPUT_BOOL_SIZE(0);\n\
+extern const std::size_t OPLCGLUE_INPUT_BOOL_SIZE(0);\n\
 GlueBoolGroup oplc_bool_input_vars[] = {\n\
 };\n\n"
 #define EMPTY_OUTPUT_BOOL "/// Size of the array of output boolean variables.\n\
-extern std::size_t const OPLCGLUE_OUTPUT_BOOL_SIZE(0);\n\
+extern const std::size_t OPLCGLUE_OUTPUT_BOOL_SIZE(0);\n\
 GlueBoolGroup oplc_bool_output_vars[] = {\n\
 };\n\n"
 #define GLUE_PREFIX "/// The size of the array of glue variables.\n"
@@ -79,7 +79,7 @@ SCENARIO("", "") {
                 "GlueBoolGroup ___IG0 { .index=0, .values={ __IX0, nullptr, nullptr, nullptr, nullptr, nullptr, nullptr, nullptr, } };\n"
                 "GlueBoolGroup* __IG0(&___IG0);\n"
                 GLUE_PREFIX
-                "extern std::size_t const OPLCGLUE_GLUE_SIZE(1);\n"
+                "extern const std::size_t OPLCGLUE_GLUE_SIZE(1);\n"
                 "/// The packed glue variables.\n"
                 "extern const GlueVariable oplc_glue_vars[] = {\n"
                 "    { IECLDT_IN, IECLST_BIT, 0, 0, IECVT_BOOL,  __IG0 },\n"
@@ -94,7 +94,7 @@ SCENARIO("", "") {
                 "GlueBoolGroup ___QG0 { .index=0, .values={ __QX0, nullptr, nullptr, nullptr, nullptr, nullptr, nullptr, nullptr, } };\n"
                 "GlueBoolGroup* __QG0(&___QG0);\n"
                 GLUE_PREFIX
-                "extern std::size_t const OPLCGLUE_GLUE_SIZE(1);\n"
+                "extern const std::size_t OPLCGLUE_GLUE_SIZE(1);\n"
                 "/// The packed glue variables.\n"
                 "extern const GlueVariable oplc_glue_vars[] = {\n"
                 "    { IECLDT_OUT, IECLST_BIT, 0, 0, IECVT_BOOL,  __QG0 },\n"
@@ -111,7 +111,7 @@ SCENARIO("", "") {
                 "GlueBoolGroup ___QG1 { .index=1, .values={ nullptr, nullptr, nullptr, __QX1_3, nullptr, nullptr, nullptr, nullptr, } };\n"
                 "GlueBoolGroup* __QG1(&___QG1);\n"
                 GLUE_PREFIX
-                "extern std::size_t const OPLCGLUE_GLUE_SIZE(2);\n"
+                "extern const std::size_t OPLCGLUE_GLUE_SIZE(2);\n"
                 "/// The packed glue variables.\n"
                 "extern const GlueVariable oplc_glue_vars[] = {\n"
                 "    { IECLDT_OUT, IECLST_BIT, 0, 0, IECVT_BOOL,  __QG0 },\n"
@@ -124,7 +124,7 @@ SCENARIO("", "") {
             std::stringstream input_stream("__LOCATED_VAR(BYTE,__IB0,I,B,0)");
             generate_body(input_stream, output_stream, digest);
             const char* expected = PREFIX"\tbyte_input[0] = __IB0;\n" POSTFIX GLUE_PREFIX
-                "extern std::size_t const OPLCGLUE_GLUE_SIZE(1);\n"
+                "extern const std::size_t OPLCGLUE_GLUE_SIZE(1);\n"
                 "/// The packed glue variables.\n"
                 "extern const GlueVariable oplc_glue_vars[] = {\n"
                 "    { IECLDT_IN, IECLST_BYTE, 0, 0, IECVT_BYTE,  __IB0 },\n"
@@ -136,7 +136,7 @@ SCENARIO("", "") {
             std::stringstream input_stream("__LOCATED_VAR(SINT,__IB1,I,B,1)");
             generate_body(input_stream, output_stream, digest);
             const char* expected = PREFIX "\tbyte_input[1] = __IB1;\n" POSTFIX GLUE_PREFIX
-                "extern std::size_t const OPLCGLUE_GLUE_SIZE(1);\n"
+                "extern const std::size_t OPLCGLUE_GLUE_SIZE(1);\n"
                 "/// The packed glue variables.\n"
                 "extern const GlueVariable oplc_glue_vars[] = {\n"
                 "    { IECLDT_IN, IECLST_BYTE, 1, 0, IECVT_SINT,  __IB1 },\n"
@@ -148,7 +148,7 @@ SCENARIO("", "") {
             std::stringstream input_stream("__LOCATED_VAR(SINT,__QB1,Q,B,1)");
             generate_body(input_stream, output_stream, digest);
             const char* expected = PREFIX "\tbyte_output[1] = __QB1;\n" POSTFIX GLUE_PREFIX
-                "extern std::size_t const OPLCGLUE_GLUE_SIZE(1);\n"
+                "extern const std::size_t OPLCGLUE_GLUE_SIZE(1);\n"
                 "/// The packed glue variables.\n"
                 "extern const GlueVariable oplc_glue_vars[] = {\n"
                 "    { IECLDT_OUT, IECLST_BYTE, 1, 0, IECVT_SINT,  __QB1 },\n"
@@ -160,7 +160,7 @@ SCENARIO("", "") {
             std::stringstream input_stream("__LOCATED_VAR(USINT,__IB2,I,B,2)");
             generate_body(input_stream, output_stream, digest);
             const char* expected = PREFIX "\tbyte_input[2] = __IB2;\n" POSTFIX GLUE_PREFIX
-                "extern std::size_t const OPLCGLUE_GLUE_SIZE(1);\n"
+                "extern const std::size_t OPLCGLUE_GLUE_SIZE(1);\n"
                 "/// The packed glue variables.\n"
                 "extern const GlueVariable oplc_glue_vars[] = {\n"
                 "    { IECLDT_IN, IECLST_BYTE, 2, 0, IECVT_USINT,  __IB2 },\n"
@@ -172,7 +172,7 @@ SCENARIO("", "") {
             std::stringstream input_stream("__LOCATED_VAR(WORD,__IW0,I,W,0)");
             generate_body(input_stream, output_stream, digest);
             const char* expected = PREFIX "\tint_input[0] = __IW0;\n" POSTFIX GLUE_PREFIX
-                "extern std::size_t const OPLCGLUE_GLUE_SIZE(1);\n"
+                "extern const std::size_t OPLCGLUE_GLUE_SIZE(1);\n"
                 "/// The packed glue variables.\n"
                 "extern const GlueVariable oplc_glue_vars[] = {\n"
                 "    { IECLDT_IN, IECLST_WORD, 0, 0, IECVT_WORD,  __IW0 },\n"
@@ -184,7 +184,7 @@ SCENARIO("", "") {
             std::stringstream input_stream("__LOCATED_VAR(WORD,__QW0,Q,W,0)");
             generate_body(input_stream, output_stream, digest);
             const char* expected = PREFIX "\tint_output[0] = __QW0;\n" POSTFIX GLUE_PREFIX
-                "extern std::size_t const OPLCGLUE_GLUE_SIZE(1);\n"
+                "extern const std::size_t OPLCGLUE_GLUE_SIZE(1);\n"
                 "/// The packed glue variables.\n"
                 "extern const GlueVariable oplc_glue_vars[] = {\n"
                 "    { IECLDT_OUT, IECLST_WORD, 0, 0, IECVT_WORD,  __QW0 },\n"
@@ -196,7 +196,7 @@ SCENARIO("", "") {
             std::stringstream input_stream("__LOCATED_VAR(INT,__IW1,I,W,1)");
             generate_body(input_stream, output_stream, digest);
             const char* expected = PREFIX "\tint_input[1] = __IW1;\n" POSTFIX GLUE_PREFIX
-                "extern std::size_t const OPLCGLUE_GLUE_SIZE(1);\n"
+                "extern const std::size_t OPLCGLUE_GLUE_SIZE(1);\n"
                 "/// The packed glue variables.\n"
                 "extern const GlueVariable oplc_glue_vars[] = {\n"
                 "    { IECLDT_IN, IECLST_WORD, 1, 0, IECVT_INT,  __IW1 },\n"
@@ -209,7 +209,7 @@ SCENARIO("", "") {
             generate_body(input_stream, output_stream, digest);
 
             const char* expected = PREFIX "\tint_input[2] = __IW2;\n" POSTFIX GLUE_PREFIX
-                "extern std::size_t const OPLCGLUE_GLUE_SIZE(1);\n"
+                "extern const std::size_t OPLCGLUE_GLUE_SIZE(1);\n"
                 "/// The packed glue variables.\n"
                 "extern const GlueVariable oplc_glue_vars[] = {\n"
                 "    { IECLDT_IN, IECLST_WORD, 2, 0, IECVT_UINT,  __IW2 },\n"
@@ -223,7 +223,7 @@ SCENARIO("", "") {
 
             // Note that the type-separate glue does not support REAL types
             const char* expected = PREFIX POSTFIX GLUE_PREFIX
-                "extern std::size_t const OPLCGLUE_GLUE_SIZE(2);\n"
+                "extern const std::size_t OPLCGLUE_GLUE_SIZE(2);\n"
                 "/// The packed glue variables.\n"
                 "extern const GlueVariable oplc_glue_vars[] = {\n"
                 "    { IECLDT_IN, IECLST_DOUBLEWORD, 0, 0, IECVT_REAL,  __ID0 },\n"
@@ -236,7 +236,7 @@ SCENARIO("", "") {
             std::stringstream input_stream("__LOCATED_VAR(INT,__MW2,M,W,2)");
             generate_body(input_stream, output_stream, digest);
             const char* expected = PREFIX "\tint_memory[2] = __MW2;\n" POSTFIX GLUE_PREFIX
-                "extern std::size_t const OPLCGLUE_GLUE_SIZE(1);\n"
+                "extern const std::size_t OPLCGLUE_GLUE_SIZE(1);\n"
                 "/// The packed glue variables.\n"
                 "extern const GlueVariable oplc_glue_vars[] = {\n"
                 "    { IECLDT_MEM, IECLST_WORD, 2, 0, IECVT_INT,  __MW2 },\n"
@@ -248,7 +248,7 @@ SCENARIO("", "") {
             std::stringstream input_stream("__LOCATED_VAR(DWORD,__MD2,M,D,2)");
             generate_body(input_stream, output_stream, digest);
             const char* expected = PREFIX "\tdint_memory[2] = (IEC_DINT *)__MD2;\n" POSTFIX GLUE_PREFIX
-                "extern std::size_t const OPLCGLUE_GLUE_SIZE(1);\n"
+                "extern const std::size_t OPLCGLUE_GLUE_SIZE(1);\n"
                 "/// The packed glue variables.\n"
                 "extern const GlueVariable oplc_glue_vars[] = {\n"
                 "    { IECLDT_MEM, IECLST_DOUBLEWORD, 2, 0, IECVT_DWORD,  __MD2 },\n"
@@ -260,7 +260,7 @@ SCENARIO("", "") {
             std::stringstream input_stream("__LOCATED_VAR(LINT,__ML1,M,L,1)");
             generate_body(input_stream, output_stream, digest);
             const char* expected = PREFIX "\tlint_memory[1] = (IEC_LINT *)__ML1;\n" POSTFIX GLUE_PREFIX
-                "extern std::size_t const OPLCGLUE_GLUE_SIZE(1);\n"
+                "extern const std::size_t OPLCGLUE_GLUE_SIZE(1);\n"
                 "/// The packed glue variables.\n"
                 "extern const GlueVariable oplc_glue_vars[] = {\n"
                 "    { IECLDT_MEM, IECLST_LONGWORD, 1, 0, IECVT_LINT,  __ML1 },\n"
@@ -272,7 +272,7 @@ SCENARIO("", "") {
             std::stringstream input_stream("__LOCATED_VAR(LINT,__ML1024,M,L,1024)");
             generate_body(input_stream, output_stream, digest);
             const char* expected = PREFIX "\tspecial_functions[0] = (IEC_LINT *)__ML1024;\n" POSTFIX GLUE_PREFIX
-                "extern std::size_t const OPLCGLUE_GLUE_SIZE(1);\n"
+                "extern const std::size_t OPLCGLUE_GLUE_SIZE(1);\n"
                 "/// The packed glue variables.\n"
                 "extern const GlueVariable oplc_glue_vars[] = {\n"
                 "    { IECLDT_MEM, IECLST_LONGWORD, 1024, 0, IECVT_LINT,  __ML1024 },\n"
